now satellite has a vector of ephemerides

+++ include/multirotor_sim/gtime.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 
 #include "multirotor_sim/datetime.h"
 
@@ -6,11 +7,11 @@ class DateTime;
 class GTime
 {
 public:
-    int week;
+    int64_t week;
     double tow_sec;
 
     GTime();
-    GTime(int week, double tow_sec);
+    GTime(int64_t week, double tow_sec);
     GTime(const DateTime& t);
     GTime& operator= (const DateTime& t);
     GTime operator -(const GTime& gt2) const;
